Tests: Refactor util and constant into folders
