+++ tests/src/footer.spec.js
@@ -1,5 +1,5 @@
 import { expect, test } from '@playwright/test';
-import { handleConsentPopup, waitFor } from './util';
+import { handleConsentPopup, waitFor } from './utils';
 
 test.describe('Smoke test for footer', () => {
   test.beforeEach(async ({ page }) => {
